Button in typescript
