+++ packages/button/package.json
@@ -13,7 +13,7 @@
   "author": "",
   "license": "Apache-2.0",
   "dependencies": {
-    "@material/button": "^0.35.0",
+    "@material/button": "^0.39.3",
     "@material/mwc-base": "^0.2.0",
     "@material/mwc-icon": "^0.2.0",
     "@material/mwc-ripple": "^0.2.0",
